+++ src/Helpers/Providers/Traits/HasFollowers.php
@@ -7,19 +7,17 @@ use seregazhuk\PinterestBot\Helpers\Pagination;
 
 trait HasFollowers
 {
-    use ProviderTrait;
-
     /**
      * @param array  $data
      * @param string $resourceUrl
-     * @param int    $batchesLimit
+     * @param int $limit
      *
      * @return Iterator
      */
-    public function getFollowData($data, $resourceUrl, $batchesLimit = 0)
+    public function getFollowData($data, $resourceUrl, $limit = 0)
     {
         $requestData = array_merge([$data, $resourceUrl]);
 
-        return (new Pagination($this))->paginate('getPaginatedData', $requestData, $batchesLimit);
+        return (new Pagination($this))->paginateOver('getPaginatedData', $requestData, $limit);
     }
 }